requirements.txt
kubernetes
pyyaml 
boto3
docker

